+++ CHANGES.md
@@ -146,6 +146,9 @@ profile. This started with version 0.26.0.
 - Fix a crash where `type%e nonrec t = t` was formatted as `type nonrec%e t = t`,
   which is invalid syntax. (#2712, @EmileTrotignon)
 
+- \* The formatting of `f (fun` is now consistent with `f ~a:(fun`.
+  (#2706, @EmileTrotignon)
+
 - Fix commandline parsing being quadratic in the number of arguments 
   (#2724, @let-def)